requirements.txt
Flask==3.1.0
Flask_Cors==5.0.0
numpy==1.26.4
python-dotenv==1.0.1
Requests==2.32.3
sentence_transformers==3.4.1
torch==2.6.0+cpu --extra-index-url https://download.pytorch.org/whl/cpu
